Update Licence field in package.json to match LICENSE file
In the `package.json`, this project had `"license": "New Relic"`, which is not a valid license type. I am working on a mobile app for an enterprise client that does dependency analysis to ensure all of our dependencies are permissable for use in their organization, and the tool flagged this project due to the unrecognized license.

This PR just updates the field to be "Apache-2.0" to match the license specified in the [LICENSE](./LICENSE) file.